Add Bionic
This is mostly a copy-paste from the original Bionic repo.  I added the
bionic package at the top level (`bionic`), the tests in
`tests/test_bionic` and a very simple example in
`examples/bionic/basic_workflow.py`.

There are still a lot of rough edges -- I was hoping to spend more time
rearchitecting things before sharing this.  Some parts of the API are a
bit more annoying to use than in the design doc.  `bionic/TODO.md` has a
roughly-ordered list of things I plan to fix soon.