Default to persistent caching
Now all derived resources are automatically cached to disk unless you
explicitly mark them with `@bn.persist(False)`.

This is the first case where Bionic needs to instantiate some core
infrastructure (the PersistentCache object) before it can run the rest
of the workflow.  We handle it by having a separate "bootstrap" mode of
computation that doesn't allow caching, multiplicity, etc.  The
advantage of using this mode, rather than exposing a separate
configuration API, is that we can still use dependency injection.  E.g.,
the user can configure the cache in one of two ways:

    # Change the input to the PersistentCache object.
    builder.set('core__persistent_cache__dir_name', './mydata')
    # Or, just provide a different implementation altogether.
    builder.set('core__persistent_cache', MyOwnCacheClass())

This is also how I'd expect us to provide executors (multiprocessing,
distributed, etc.) and so on.

BTW, at some point I'd like to make the double underscore interchangeable
with a period, so that we can do

    builder.set('core.persistent_cache.dir_name', ...)
    ...
    def core__persistent_cache(core__persistent_cache__dir_name):
        ...