Add name to FlowBuilder/Flow objects
It seems like a good idea for flows to have names, if only to avoid
their persisted data from overlapping.  Also, I think this may be useful
for making a reload helper.

Providing a name is now mandatory when you create a FlowBuilder, which
doesn't seem like too much to ask.