Use persistent data structures for FlowState
Previously the FlowBuilder and Flow classes accessed their state through
a handle object, which allowed us to switch between mutable and
immutable access while avoiding making a copy on every update.  However,
the handle code was confusing and hard to reason about.

Now the FlowState object uses persistent data structures, which means
it's immutable but we can efficiently create modified copies.  This
eliminates the need for the handle interface.  It also simplifies a lot
of our FlowBuilder code, because we don't have to worry about
encountering a failure partway through a series of update operations;
instead we do the updates on a series of copies, and then adopt the
final copy at the end.

Finally, I think this may help with reloading modules later.