Add Flow.reloading()
This magically reloads the modules used by the flow; it should make it
easier to manage situations where a flow is defined in a module and then
used from a notebook.  Reloading only works in certain circumstances,
but hopefully it will make iteration and testing easier in the simple
80% case.  We will see!

I considered some other approaches but they were either very cumbersome
for the user or seemed like they would silently fail in some edge cases.