Fix reload in Python 2
`importlib.reload()` doesn't exist in Python 2, so we need to use `six`
to get a consistent interface.

I spent a while trying to make a test for this, but I couldn't get it
working. There was some kind of caching going on in the filesystem or in
CPython that I couldn't figure out.  It does seem to work in a notebook
though.