Avoid creating files starting with a `.`
Currently we clean filenames by replacing bad characters with `.`s,
which is problematic because Unix treats files starting with `.`s as
"hidden" files.  To avoid this I'm adding a `_` in front of any filename
that would otherwise start with a `.`.  (I considered just using `_` for
the cleaning everywhere, but `.` feels a bit cleaner-looking somehow.)