Update DAG visualization
I made the following changes to the GraphViz DAG rendering:
- same-resource nodes are clustered together
- nodes are colored, again based on resource
- by default, layout is left-to-right instead of top-to-bottom
- by default, edges are straight lines instead of curved lines

Overall I think this looks a little less elegant when the graph is
simple (one node per resource), but much better when the graph is
complex (multiple nodes per resource).

In the future I think it would make sens to color nodes based on cache
status (cached in memory, on disk, or not at all) instead of resource,
but that would require a bit more refactoring.