Add wrapper to fix "unhashable type" errors
When you call something like `flow.get.model('series')`, Bionic tries to
return a Series whose values are the models and whose index contains all
the root ancestor values for those models.  This is a problem if any of
those ancestors has an unhashable type, like a dict of hyperparameters.

I've fixed this by putting a hashable wrapper around all the index
values.  This makes it a little more annoying to access the index items,
but it avoids confusing "unhashable type" exceptions.

I think a better solution might be to require that if you have multiple
values for a resource, either they need to be hashable *or* you need to
provide a list of hashable labels *or* you need to provide a function
that performs the labeling.  But for now I'd like to get rid of the
exception, because it can come up pretty easily in the ML workflow
example.