Refactor caching to support GCS
Everything is still cached to local disk, but there is now the option of
also having them copied to and from GCS.  If GCS caching is enabled,
Bionic will read from the GCS cache when a local copy is not available,
and always copy any accessed data to any cache that doesn't have it.

This also moves the caching mechanism to a content-addressed system
where nothing needs to get overwritten, which means we don't need to use
symlinks to do atomic writes.  This makes the cache directory harder for
humans to browse, but it also makes it possible to implement it on top
of things like GCS that don't support symlinks.