Be explicit about format when saving images
If you don't specify a file format when saving a Pillow Image, it will
infer it from the filename.  Normally this works fine, but in some cases
it can get confused (like when you pass it a file object created by
PyFilesystem in Python 3).  It makes more sense to be explicit here
anyway.