Store temp files in bndata/
Background: All temp files are written to a temporary directory first
and then moved to their permanent cache locations and replaced with
temporary symlinks.  Previously this temp directory was in a
system-default location -- however, this caused problems when that
location was on a different filesystem from the permanent cache
directory (`bndata/`), because you can't have symlinks across
filesystems.

With this commit, we write the temp files in a subdirectory of
`bndata/`, so everything should stay on the same filesystem.