Remove dsutil module
This was a collection of personal utilities that I included when Bionic
was just a prototype.  There's no real reason to have it in this project
now, and removing it allows us to drop the `ipython` dependency.

I also moved the `jupyter` dependency to `[dev]`, since we only need it
for building the docs.