Refactor caching to avoid loading from temp files
The original cache behavior was to do all serializing/deserializing
to/from a temporary directory, and then move files between there and the
different cache locations.  The problem is that some objects
-- specifically Dask DataFrames -- need to be backed by a long-lived
file, not a temporary one.  This commit refactors the cache to have all
loading happen directly from the local disk cache, where we expect files
to live indefinitely.  This should make it possible to implement a
protocol for Dask frames.