Add assisted versioning mode
This introduces two new versioning modes: "assist" and "auto".

Assist: Use each function's bytecode to attempt to detect when the user
has forgotten to update the function's version -- if so, throw an
exception.

Auto: Implicitly update each function's version when its bytecode
changes.

These modes are opt-in; they need to be explicitly enabled.  This is
partly because they introduce a large amount of new complexity to the
code base, so we may want to roll them back if they're not helpful and
reliable.