Raise ValueError when writing DataFrame with duplicate columns
Previously, the call to self._check_dtypes within
ParquetDataFrameProtocol#write would fail with duplicate columns because
it expects df[col] to always be a series.  If the col in question
appears more than once, df[col] produces a DataFrame.  We do not support
writing DataFrames with duplicate columns.  This commit raises
a ValueError if it encounters this scenario when writing.