Fix failure when persisting None values
The new caching code was failing whenever we tried to persist a None
value; this commit should fix it.

The caching code was getting confused by the difference between "we have
a value which is None" and "we don't have a value at all".  I took a
quick look for other situations where we use this pattern but couldn't
find any -- most of the Bionic code doesn't care about the content of
what's being persisted.