Import PIL.Image directly
There were some parts of the code where we were importing `PIL` and then
accessing `PIL.Image`, which only works if you've previously imported
`PIL.Image` directly.  Everything tended to work because of the order in
which things were imported, but it was possible to get things to happen
in the wrong order, which would throw a confusing exception.