Handle older caching filename convention
In Bionic 0.5.7 we mistakenly updated the naming convention for cached
files in a non-backwards-compatible way.  (We switched from naming all
files `value.EXTENSION` to naming them `ENTITY_NAME.EXTENSION`.)  That
means anyone using the new code will be unable to load files saved with
the old version.

This commit updates the loading code to handle both the old and new
extension, which should fix this problem.