Introduce helper for reformatting error messages
This `oneline` helper function allows us to write our error messages as
multiline strings but have them come out as a single line, which is
easier to manage than several lines of separate strings (some of which
are f-strings and some not).  In particular, it's now possible to use
automatic tools Vim's `gq` to wrap lines in these messages.