Remove custom matplotlib initialization
This removes the Matplotlib initialization code which replaces the
`MacOSX` backend with the `TkAgg` backend.  This used to be helpful
because the `MacOSX` backend didn't work for some Python builds.
However, as of Matplotlib 3.1, the `MacOSX` backend should work for all
builds.  This makes it a better default than `TkAgg`, which has a
dependency on Tcl/Tk.

Accordingly, I have also set the minimum version of Matplotlib to 3.1.