Add `@protocol.path`
This introduces a protocol for pathlib Paths referring to local files,
which allows entities to create files directly and have them be managed
by Bionic's tiered caching.

In the future I'd like to have a more elegant way of working with files,
but that requires a lot more infrastructure work.