Introduce descriptors into Bionic internals
This moves a lot of Bionic's internal code to using DescriptorNodes, or
"dnodes", instead of simple entity names. Currently dnodes are just
wrappers for entity names, but in the future they will be able to
represent more complex structures.

There should be no functional changes in this commit, except that we're
updating the cache schema version.

Finally, this commit introduces a dependency on the
[attrs](https://www.attrs.org/en/stable/) package. I think this is a
cleaner way to create customized data classes than `namedtuple`, and I'd
be interested in converting some of the classes in `datatypes.py` to use
`attrs` as well.