Update docs to support ZSH
Replaced all instances of

    pip install bionic[SOMETHING]

with

    pip install 'bionic[SOMETHING]'

The former doesn't work on Zsh, since that shell assigns a special
meaning to square brackets. Having to add the quotes is kind of
annoying, but now Zsh is the default shell on new Macbooks and it
generates a cryptic error if you don't add the quotes, so it seems
worthwhile to try to help newcomers avoid this.