Remove bad assertions in TaskState completion
These assertions would case derivation to crash if we tried to compute
an entity twice and the entity function had an internal exception the
first time. For example:

    import bionic as bn

    builder = bn.FlowBuilder('test')

    @builder
    def y():
        return z

    flow = builder.build()

    if __name__ == '__main__':
        import traceback
        try:
            print(flow.get('y'))
        except Exception as e:
            # Throws correct exception.
            traceback.print_exc()
        try:
            print(flow.get('y'))
        except Exception as e:
            # Throws incorrect assertion error.
            traceback.print_exc()

These assertions were left over from when different parts of the
TaskState completion process were run independently; now that they're
all part of a single function, we need to handle the case where the
completion terminated halfway.

I'll add a test case in the next commit.