Create TaskStates already initialized
Instead of creating all the TaskStates in one pass and then initializing
them in a second pass, we now initialize each state as we create it.
This unblocks the ability to incrementally add more TaskStates over
time, rather than creating them all at once.

However, to do this we have to remove the bi-directional links between
states, since it's impossible to maintain them without going back and
editing existing states. So, I've removed the `children` field of
`TaskState`. This field is mainly useful for keeping track of which
states are waiting for the current one to finish; that logic has been
moved into a BlockageTracker class which is only used during
computation. This is part of a continued project of separating out the
per-flow-instance and per-get-call parts of EntityDeriver and TaskState.