Log each task key exactly once per `get()`
When you call `Flow.get()`, it emits a log message for each relevant
task key as it loads or computes the value for that key. It used to be
the case that each key would get logged exactly once, so if a key was
computed and then later loaded from a cache, it would only log the
computation. This made it easy to see what was going on.

Unfortunately, when we updated the cache invalidation behavior, we lost
this property (and, perhaps more unfortunately, there are no tests for
it). This commit restores that property by introducing a TaskKeyLogger
class. It also adds a test.

Passing this logger object around is kind of annoying, but I think we
may be able to eliminate this in future refactoring. I'm looking into
factoring out a separate EntityDerivation class, which contains all the
state specific to a single `Flow.get()` call. This includes both the
task key logging and the tracking of which task states are blocked by
which.