Compute task/descriptor info incrementally
This commit makes it possible to compute task metadata incrementally and
lazily, instead of doing it all at once for all tasks during
initialization. This will be important later when we introduce more
complex descriptors, since the number of possible descriptors will be
infinite.

We still pre-compute all the "base" descriptors ahead of time in
order to surface errors as early as possible.

I introduced a DescriptorInfo class to hold per-descriptor metadata.
Eventually descriptors will be 1:1 with tasks, so we'll want to refactor
this and TaskState.

Finally, I added docstrings for all the private methods in
EntityDeriver.