Remove extension param from protocol.read()
Currently all protocols have a `read()` method which accepts both a
path and a file extension. The file extension argument is a bit weird:
it's never used except by the `CombinedProtocol`, which needs it to
figure out which sub-protocol to use. (The extension doesn't necessarily
correspond to `path.suffix`, because it can contain dots, like
`pq.dask`.)

This commit removes the extension argument, but allows protocols to
implement an alternative method, `read_with_extension`, if they need it.
This simplifies the other 10 or so protocols, and also brings the
interface back into sync with our documentation
[here](https://bionic.readthedocs.io/en/latest/api/protocols.html#custom-protocols).