Fix bug breaking the use of fixed dataframes
In commit 3b86ba0fdbc83a4f95bf14b84e3d030365313bcf, I added some code
that applied `==` to user-provided values. This was a bad idea because some
classes (like pandas.DataFrame) override `==` and return a value that
can't be interpreted as a boolean. Thankfully this caused our
documentation build to fail.

I changed the way we do this comparison and added a test to catch this.