Refactor decoration accumulation
This is the first step towards addressing #127.

This changes the way Bionic decorators are implemented. Previously each
decorator worked by accepting either a function or a provider, and
wrapping it in a new provider. Now they accept a function and return it
after modifying a special `bionic_decorator_accumulator` attribute.
This is conceptually cleaner (no more `func_or_provider`) and introduces
the `DecorationAccumulator` class described in the issue linked above.