Mark fixed entities as _can_persist
Previously fixed entities had _can_persist=False, which sort of makes
sense since we don't actually want to persist them. However, since we do
expect fixed entities to be serializable, it's not really true that we
*can't* persist them. With this change, we now mark them as
`_can_persist=True`, but update `TaskState.should_persist` to be
False for fixed entities, so all behavior stays the same.

The upshot is that now all entities default to being persistable. This
makes it easier to break out a separate immutable entity definition
class, since the persistence properties don't change when you convert a
derived entity to a fixed one.