Support logging for parallel execution
This change uses a queue to send logs from the subprocesses back to
main process. The queue is installed as a part of logging handlers
`ProcessPoolHandler` in the subprocesses and a receiver
`LoggingReceiver` recevies the messages from the queue in the main
process and propagates the records to appropriate loggers.