Add flake8-fixme
This will cause flake8 to fail if any of our code includes a "FIXME". By
convention we use this string to flag things that need to be changed
before merging, so it's useful to check this automatically. (For things
that need to be addressed sometime farther in the future we use TODO,
which is still allowed.)