Refactor `results_by_name` to `results_by_dnode`
The point of this field is to hold multiple results for a single
task (that differ only by name). Once we have tuple descriptors, we'll
be able to remove this concept and every task will have a single result.
However, in order to implement tuple descriptors we need to support
descriptors that aren't entity names, so for now it's convenient to
organize results by descriptor.

I also made corresponding changes to `value_hashes_by_name` and
`key_for_entity_name`.

There should be no functional change.