Use tuple descriptors internally
This changes implementation of `@outputs` and `builder.adding_cases` to
use tuple descriptors instead of using providers that return multiple
values. This has no functional changes, with the exception that tuples
nodes now appear in the DAG visualization as their own nodes.  (The
logging for these nodes is probably also slightly different.)

One big benefit of this change is that now every task has exactly one
key and returns exactly one value, which means we can substantially
simplify a lot of code in a future refactor.