Use a more interesting graph for dagviz tests
The original tests used a graph with only one node; now we use a larger
graph that includes a multi-output task, multiplicity, and a gather
operation. We also test more properties of the pydot output.

These tests will be even more important when we start using tuple
descriptors, since we'll want to be sure that they can be visualized
properly.