Stop using CaseKey.values
CaseKeys are sent to the subprocesses as part of a lot of objects. We
don't want values to be sent across directly to subprocesses as they
can be unpicklable or user might have defined a custom protocol for
them that we should use instead.

This change stops using values inside CaseKey. It is only used to get
the values for the multi-index when fetching a series for the user.
Instead of storing values in CaseKey, we now store them inside
providers which already takes care of CaseKeys for us, are immutable,
and aren't sent to the subprocesses.