Update parallel processing executor to use correct worker count
Bionic maintains a single instance of process pool executor that it
shares between all flow instances in the current process. If the worker
count between flows is different, that creates a problem since the
executor is shared.

Till now, Bionic used the worker count as defined by the last flow and
emitted a user warning if the worker count changed. With this change,
Bionic resizes the executor using each flows worker count before any
entity is derived. This removes the user overhead of making sure that
the worker count is correct and allows them to experiment with
different worker count as well.