Write user defined values to tmp cache
In parallel computing mode, Bionic currently passes all the user
defined values to the sub processes using IPC. This can break if the
values are not pickable which user can get around by writing a custom
protocol.

This change writes all the user provided values to the cache that can
be read by the subprocesses. This allows us to take advantage of
protocol serialization and not worry about breaking our user contract.