Make ProcessManager global
ProcessManager was defined inside ExternalProcessLoggingManager.__init__
which makes ProcessManager class unpickable. We noticed this problem
when running parallel computing tests on MacOS with python version 3.8.
In python 3.8, multiprocessing uses spawn instead of fork as the default
mode which requires everything passed to be pickable.

This change makes ProcessManager a global class so it can be pickled and
parallel computing works for MacOS on python 3.8.