Reset tmp cache after each run
The temporary values have a life span of a get call. As soon as the
entities are derived in `Deriver._compute_result_group_for_dnode`, we
reset the tmp cache.

As a fail-safe, we also attempt to reset the tmp cache before deriving
entities. This protects against cases where the execution failed while
deriving values and leaves the cache and more importantly, task states
in a bad state.