Allow internal entities to depend on persistable user values
Instead of throwing an exception that users can't use persistable
values as a dependent for internal entities, we now allow this by
not persisting those user provided values and raising a warning.

This keeps Bionic backwards compatible. User provided values are
persisted by default but they didn't before. If we disallow this
behavior, this will be a breaking change as users might be using their
own entities to generate internal entities.