Add listeners to logging receiver
Bionic uses a logging receiver to receive logs from the subprocesses
and log them in the main process. To write the logs in main process, it
runs a loop in thread that writes whenever it finds an object in the
logging queue used to send over logs from the subprocesses.

Instead of continuously looping and checking the logging queue in that
thread, we now use a count of listeners on the logging receiver. The
loop checks the queue only if there are any active listeners.

Aside from minor computing save, this also fixes the issue where when
the process ends, receiver thread throws an exception trying to access
queue object managed by the manager that stops before killing this
daemon thread. https://github.com/square/bionic/issues/161