Reuse flows in fuzz tests
Fuzz tests were creating a new flow for every get operation which was
costly. This change keeps a flow object and updates it any time an
entity is added or updated. This way, we reduce the number of entities
to be completed, which is specially useful for parallel execution as
each completion takes place in a subprocess. We end up spending a lot
of time in IPC for basically just loading cached entry in cache
accessor from disk cache otherwise.

This change reduced the time taken by parallel fuzz tests by 33% on my
machine on python 3.8.