Use subprocesses only for actual computation
Bionic completes task states in parallel when parallel execution is
turned on. But task states that are already computed and cached on disk
might not be marked complete. For such cases, we end up completing
those task states in the subprocesses, which barely does anything and
adds IPC cost to the execution.

Bionic now only uses subprocesses for the actual compute step, not for
the full completion step. Computing is what we essentially want to
parallelize anyway. This helps us save IPC cost for the cases described
above.

As a bonus, our total time for parallel fuzz tests also went down
by around 15-20% again.