Allow uncached entities
Bionic did not allow entities to be uncached. All the entities had to
be either memoized or persisted or both. This change removes that
restriction and allows entities to be uncached. The entities are now
computed every time they are needed.

This also simplifies some of the `TaskState` code by removing
`TaskState.is_complete` in favor of `TaskState.is_cached` property.