Refresh cache state with each `get` call
The goal here is to gracefully handle the situation where cached files
are deleted in between one `Flow.get` and another. To do this, we
re-sync the cache state on each TaskState before doing any processing on
it.

This will be important when we expose a cache entry deletion API.