Gracefully handle metadata files with no artifact
Now if we load a metadata file whose artifact URL doesn't exist, we
delete the metadata and act like it never existed. This makes it safe to
delete artifact files without also deleting their metadata. In
particular, this fixes a bug in the cache API: if two metadata files
point to the same artifact, the deleting the entry for one of them would
make the other invalid. Now we can handle this situation gracefully.

I've also swapped the order in which cache entry URLs are deleted: we
now delete the artifact first, and then the metadata. Now that it's safe
to have broken URLs in metadata, it's better to avoid orphaning the
artifact.