Format logs before sending them to the main process
Bionic sends all the log records to the main process and lets the main
process log them. In this case, the log formatting happens in the main
process, which requires us to send the entire argument list of the
logs. The communication between processes requires every argument to be
pickleable using cloudpickle, which seems unreasonable to enforce.

To avoid breaking logging, we eagerly format the log message the way
logging does (check LogRecord.getMessage). Although the logging works,
this has a side effect of changing the template and arguments for the
log in case users intercepts the LogRecord in their custom handlers.