Fix handling of entity ordering in add_case
The `FlowBuilder.add_case` method accepts multiple name-value pairs;
these can be given in any order, and are automatically permuted to get
them into a canonical order. Unfortunately, we had a bug where we
instead applied the inverse of the correct permutation.

As it happened, our test for this logic only required permutations with
a single swap -- which meant there was no difference between a correct
permutation and its inverse.

Oh, and that test case was also missing the word `assert` in all its
checks.