Move to stack-based execution
We moved to recursion-based execution inside the subprocesses whenever
any dependencies were unserializable, mostly because it was easier to
send `TaskState.compute()` to the subprocesses.

But now that we send FlowCpmpletionRunner to the subprocesses instead, we
can get rid of the dual execution strategy and move the subprocesses to
stack-based execution as well.

Note that we haven't completely gotten rid of recursion-based execution
for entities that are not cached. They are always computed and not
cached by definition.