Compute memoize-only entries using stack-based execution
Bionic computes entries using a stack-based execution logic. But for
some memoize-only entries, it uses a recursion-based approach. We
change the logic to compute such entities using the unified stack-based
approach.

It does so by adding a `SKIPPED` EntryStage to differentiate between
skipped and completed entries. We also track whether entries would be
needed in memory and computing memoized-only entries when they are.