Compute all entries using stack-based execution
This change is the last part of the series of changes that make all
entry computation stack-based. Only non-cacheable entries before this
change were computed recursively. Now, we store their results on the
entries themselves. We are technically caching them; it's just that the
cached values are short-lived and discarded with the completion runner.

To make this change, compute and blocking deps methods are now on
TaskRunnerEntry. By doing so, we also simplified the dep methods
substantially.