Fix bug loading multifile artifacts from GCS
We had a bug where Bionic would fail to recognize existing multifile
artifacts in GCS and would just recompute them instead. This was caught
by our GCS tests, but these don't run in CI and apparently I failed to
run them since introducing the bug in
462c6c3c74290f686b24ad471c6b8c003d456ea3 ("Gracefully handle metadata
files with no artifact"). Oops!

At this point I'm tempted to try implementing a simple mock of the GCS
API to allow us to catch stuff like this without manually running the
GCS tests, which are slow and can't run in CI.