Move `oneline` into `utils`, and `gcs` out
This moves the `oneline` function into `utils.misc`. This would cause a
circular dependency between `deps` and `utils` --, because `utils.gcs`
depends on `deps`, which depends one `oneline` -- so I also moved
`utils.gcs` to the top level. Now nothing in `utils` depends on any
optional dependencies, so it's safe for `deps` to depend on it.