Add follow-up tasks for tuples
This adds the concept of "follow-up tasks": when a task is finished, if
it has any "follow-up" tasks, those tasks must also be completed
immediately. This allows us to guarantee that when a tuple-returning
task is computed, all of the individual entities in the tuple are
persisted immediately.

Implementation-wise, this includes the following changes:
- Modifying TaskState creation to include these followups (which can now
introduce circular relationships between tasks)
- Modifying the TaskCompletionRunner to:
  - use a "priority stack" instead of a simple list
  - queue up follow-ups, at high priority, at the appropriate time
  - when dispatching tasks remotely, include any additional follow-ups
    that need to be computed as well