Avoid recomputing deps unnecessarily
This is a substantial refactor of TaskCompletionRunner to avoid
recomputing non-persistable entities before loading their dependents
from the cache. The problem is this: in order to load a value from the
persistent cache, we need to compute the provenance digest of each of
its dependencies, which currently requires computing them. This means
that if a task has any non-persistable dependencies, they need to be
recomputed every time we want to load it from the cache. However, it's
also the case that when tasks have non-persistable output, their
provenance digest doesn't depend on their value, so this extra
computation is not even required.

To avoid this computation, we allow these non-persistable tasks to be in
a special "primed" state, where their digest has been computed but their
actual value has not. As it happens, this state is also useful for
tracking whether these tasks are ready to run in a subprocess.

I also made the following adjustments:

- I removed the NEW and DEFERRED stages for TaskEntries. I think the
right way to think about entry stages is that an entry is COMPLETED as
soon as it has met all known requirements; however, an entry can become
un-COMPLETED when new requirements become known. This means that
DEFERRED and NEW can both be absorbed into COMPLETED.

- Previously task entries had a concept of "needed in memory" vs "needed
but not in memory". I changed this to "requested to have cached values"
vs "requested to be primed", which I think is a more useful distinction.

- The `_process_active_entry` method now contains the core of our
asynchronous logic, which has become more complex. The goal here is to
do as little work as possible to meet the requirements on an entry.

Fixes https://github.com/square/bionic/issues/111.