Remove need for `count_calls`
This makes the output of the `make_counter` fixture usable as a
decorator, so instead of writing `@count_calls(counter)`, we can just
write `@counter`.

This is a little tricky because of the way SyncManager proxies work: I
had to split the ResettingCounter class into a SimpleCounter class
(which can be proxied) and a ResettingCallCounter class (which can't,
but wraps a SimpleCounter).