Fix test failure due to Dask update
Older versions of Dask (before 2.22.0) couldn't read multi-index
dataframes from Parquet, so we have a check to detect this and throw an
error if it happens, and a test to make sure this check is working.
However, the latest version of Dask will now happily write and read
multi-index dataframes... but will sometimes return the sub-indices in
the wrong order. I don't know of a good way to catch this
non-deterministic behavior, so for now I'm just loosening the test to
accept it.

More details: https://github.com/dask/dask/issues/6484