Introduce FunctionAttributes class
I moved the `code_fingerprint` and `changes_per_run` attributes into
a small data class, which `TaskState` can use instead of keeping a
reference to the original `Provider`. This makes no real difference now,
but it gives us the option of simplifying our task stripping code,
because `FunctionAttributes` is picklable but `Provider` is not. I think
this will be useful for fixing tuple persistence.

I also added a TODO for a further refactor that would move
`FunctionAttributes` onto the `Task` itself.