Disallow Sphinx 3.2.0
This version of Sphinx fails to build our docs with the following error:

    Handler <function _process_docstring at 0x7f156ef03598> for event
    'autodoc-process-docstring' threw an exception (exception: module,
    class, method, function, traceback, frame, or code object was expected,
    got PathProtocol)

This appears to be a bug in Sphinx which should be fixed in the next
version. (See [here](https://github.com/sphinx-doc/sphinx/issues/8074)
for more details.) I've confirmed that the build works properly with
Sphinx master.