Add CodeHasher class to hash code objects
Bionic detects changes in each entity's function code and invalidates
its cache if the code changes. It does so by looking at assembly
instructions and hashing them into a cache key. But it only looks at
the function code; it doesn't detect any references to other functions.

This change adds a different way to hash the function code. The new
logic is in the CodeHasher class, which looks at the code objects
instead of assembly instructions. CodeHasher also detects changes to
only the function code, but it's done in a way that it will be easy to
extend and check the references.