Add new random flow tester
This is an alternative implementation of the random testing ideas in
`test_persistence_fuzz.py`, but focused on persistence vs
non-persistence and multi-output functions instead of versioning. The
goal is to allow us to randomly test the tuple persistence capabilities
(both serial and parallel) that I've been working on.

I basically rewrote `test_persistence_fuzz.py` because I was finding it
hard to add support for the new features above. The new implementation
models a little more of what's going on in a flow, so it shouldn't be
too hard to support what the fuzzer does as well (versioning modes, GCS,
and non-determinism) and eventually remove the old fuzzer code
altogether.

The new implementation also has a different algorithm for generating
graphs which produces more realistic and efficient flows, which means we
can test larger flows in a reasonable amount of time.

Finally, I've avoided the term "fuzzer" in this new implementation
because what we're doing doesn't actually feel like "fuzz testing". I
think "random testing" or "model-based testing" are more accurate.