Improve error msg of `@outputs` on non-seq values
The `@outputs` decorator expects to be used on functions that return a
sequence of a specific size. If we get a sequence of the wrong size, we
return a clear(-ish) message, but if we get something that's not a
sequence, we fail cryptically. This commit adds a clearer error message
for that situation.