Add tests for GC of non-memoized values
I want to enforce the property that if an entity is not memoized, then
we don't store it in memory longer than necessary. This is currently
true, but I think it will break when we activate tuple descriptors
(unless/until we add some smarter garbage collection in our execution
code).